requirements.txt
streamlit
numpy
pillow
torch
transformers
mediapipe

